requirements.txt
flask
gunicorn
flask_cors
psycopg2-binary
pyTelegramBotAPI
pybase64
flask_autoindex
wfuzz
wafw00f
bs4
colorama
requests
argparse
python-Wappalyzer
aiodns==3.0.0
aiofiles==0.7.0
aiohttp
aiomultiprocess==0.9.0
aiosqlite==0.17.0
beautifulsoup4==4.10.0
censys==2.0.8
certifi==2021.05.30
dnspython==2.1.0
fastapi==0.68.1
netaddr==0.8.0
orjson==3.6.3
pyppeteer==0.2.6
PyYAML==5.4.1
requests==2.26.0
retrying==1.3.3
setuptools==58.1.0
shodan==1.25.0
slowapi==0.1.5
spyse-python==2.2.3
starlette==0.14.2
lxml
uvicorn==0.15.0
uvloop
emailprotectionslib
Resolver
